Added new Controller
I added another controller just for the menu window and I also cleaned up all of the code in each file. I tried to put appropriate comments on all of the methods and variables that we might have trouble identifying. I also was able to make the monster battle and end screens provide the user with more information about what happened.